filter on stats

+++ Front/src/app/events/events-list/events-list.component.ts
@@ -54,17 +54,30 @@ export class EventsListComponent implements OnInit{
   }
   removeInterest(event:Event){
     console.log("")
-    this.eventService.deleteInterestByUserByEvent(this.userId,event.id).subscribe((data)=>
+    if (this.isEventInList(event.id))
     {
-
-      this.eventList[ this.eventList.findIndex((ele)=>ele.id==event.id)]=data
-      this.isEventInList(event.id)
-    }
-    );
+      this.eventService.deleteInterestByUserByEvent(this.userId,event.id).subscribe((data)=>
+      {
+        console.log( this.eventList[this.eventList.findIndex((ele)=>ele.id==event.id)])
+        console.log(data)
+        this.eventList[this.eventList.findIndex((ele)=>ele.id==event.id)]=data
     
-    console.log("Removed Interest")
+      });
+      console.log("Removed Interest")
+
+    }
+    else
+    {
+      this.addInterest(event)
+      console.log("Added Interest")
+
+    }
+    
+    
   }
   isEventInList(eventId: number): boolean {
+    if( this.myInteresetList.length==0)
+    return false;
     return this.myInteresetList.findIndex((event) => event.eventID === eventId) !== -1;
   }
